+++ docs/gallery/advanced_io/streaming.py
@@ -7,23 +7,20 @@ from the `DANDI Archive <https://dandiarchive.org/>`_. This is especially useful
 from a large NWB file stored remotely.
 '''
 
-# sphinx_gallery_thumbnail_path = 'figures/gallery_thumbnails_streaming.png'
-
 ####################
 # Streaming data from an S3 bucket requires having HDF5 installed with the ROS3 (read-only S3) driver.
-# The example code below also requires the ``requests`` package to be installed.
 # You can install HDF5 with the ROS3 driver from `conda-forge <https://conda-forge.org/>`_ using ``conda``.
 # You may first need to uninstall a currently installed version of h5py.
 #
 # .. code-block:: bash
 #
 #   pip uninstall h5py
-#   conda install -c conda-forge "h5py>=3.2" requests
+#   conda install -c conda-forge "h5py>=3.2"
 #
 
 ####################
 # Next, use the ``DandiAPIClient`` to get the S3 URL to an NWB file of interest stored in the DANDI Archive.
-# If you have not already, install the ``dandi`` package.
+# If you have not already, install the latest version of the ``dandi`` package (0.27 or larger).
 #
 # .. code-block:: bash
 #
@@ -37,14 +34,14 @@ from a large NWB file stored remotely.
 #   dandiset_id = '000006'  # ephys dataset from the Svoboda Lab
 #   filepath = 'sub-anm372795/sub-anm372795_ses-20170718.nwb'  # 450 kB file
 #   with DandiAPIClient() as client:
-#     asset = client.get_dandiset(dandiset_id, "draft").get_asset_by_path("filepath")
+#     asset = client.get_dandiset(dandiset_id, 'draft').get_asset_by_path('filepath')
 #     s3_path = asset.get_content_url(follow_redirects=1, strip_query=True)
 
 ####################
 # Finally, instantiate a :py:class:`~pynwb.NWBHDF5IO` object with the S3 URL and specify the driver as "ros3". This
-# will download metadata about the file from the S3 bucket to memory. The values of Datasets are accessed lazily,
+# will download metadata about the file from the S3 bucket to memory. The values of datasets are accessed lazily,
 # just like when reading an NWB file stored locally. So, slicing into a dataset will require additional time to
-# download the sliced data to memory.
+# download the sliced data (and only the sliced data) to memory.
 #
 # .. code-block:: python
 #
@@ -53,6 +50,7 @@ from a large NWB file stored remotely.
 #   with NWBHDF5IO(s3_path, mode='r', load_namespaces=True, driver='ros3') as io:
 #       nwbfile = io.read()
 #       print(nwbfile)
-#       print(nwbfile.acquisition['lick_times'].time_series['lick_left_times'])
 #       print(nwbfile.acquisition['lick_times'].time_series['lick_left_times'].data[:])
 #
+
+# sphinx_gallery_thumbnail_path = 'figures/gallery_thumbnails_streaming.png'